исправляет указанные ошибки
